+++ Domain/Order/ICalculatePriceStrategy.cs
@@ -2,5 +2,5 @@ namespace Data;
 
 public interface ICalculatePriceStrategy
 {
-    public decimal CalculatePrice(List<MovieTicket> tickets, bool isStudentOrder);
+    public double CalculatePrice(List<MovieTicket> tickets, bool isStudentOrder);
 }
